Naive semantic handling - adding all possible syns to query and search it best : w1 = 2, w2=0.4

+++ GUI/MainWindow.xaml.cs
@@ -49,7 +49,7 @@ namespace GUI
         DocumentRank[] rankedDocument;
         static char[] QuerrySplitters = new char[] { ' ', '\t' };
         bool searchDone = false;
-        double w1=2;
+        
 
         /// <summary>
         /// Ctor for the MainWindow
@@ -78,72 +78,72 @@ namespace GUI
         private void InitLanguages()
         {
             languages = new Dictionary<string, HashSet<string>>();
-            languages["Afrikaans"] = new HashSet<string>{ "Afrikaans" };
-            languages["Albanian"] = new HashSet<string>{ "Albanian" };
-            languages["Amharic"] = new HashSet<string>{ "Amharic" };
-            languages["Arabic"] = new HashSet<string>{ "Arabic", "Arabi" };
-            languages["Armenian"] = new HashSet<string>{ "Armenian" };
-            languages["Azeri"] = new HashSet<string>{ "Azeri" };
-            languages["Belarusian"] = new HashSet<string>{ "Belarusian" };
-            languages["Bengali"] = new HashSet<string>{ "Bengali" };
-            languages["Bulgarian"] = new HashSet<string>{ "Bulgarian" };
-            languages["Burmese"] = new HashSet<string>{ "Burmese" };
-            languages["Cambodian"] = new HashSet<string>{ "Cambodian" };
-            languages["Cantonese"] = new HashSet<string>{ "Cantonese" };
-            languages["Chinese"] = new HashSet<string>{ "Chinese" };
-            languages["Creole"] = new HashSet<string>{ "Creole" };
-            languages["Czech"] = new HashSet<string>{ "Czech" };
-            languages["Danish"] = new HashSet<string>{ "Danish" };
-            languages["Dari"] = new HashSet<string>{ "Dari" };
-            languages["Dutch"] = new HashSet<string>{ "Dutch" };
-            languages["English"] = new HashSet<string>{ "Eng", "English", "Enlgish" };
-            languages["Estonian"] = new HashSet<string>{ "Estonian" };
-            languages["Finnish"] = new HashSet<string>{ "Finnish" };
-            languages["French"] = new HashSet<string>{ "French" };
-            languages["Georgian"] = new HashSet<string>{ "Georgian" };
-            languages["German"] = new HashSet<string>{ "German" };
-            languages["Greek"] = new HashSet<string>{ "Greek" };
-            languages["Hebrew"] = new HashSet<string>{ "Hebrew" };
-            languages["Hindi"] = new HashSet<string>{ "Hindi" };
-            languages["Hungarian"] = new HashSet<string>{ "Hungarian" };
-            languages["Indonesian"] = new HashSet<string>{ "Indonesian" };
-            languages["International"] = new HashSet<string>{ "International" };
-            languages["Italian"] = new HashSet<string>{ "Italian" };
-            languages["Japanese"] = new HashSet<string>{ "Japanese" };
-            languages["Kazakh"] = new HashSet<string>{ "Kazakh" };
-            languages["Kinyarwanda"] = new HashSet<string>{ "Kinyarwanda" };
-            languages["Kirundi"] = new HashSet<string>{ "Kirundi" };
-            languages["Korean"] = new HashSet<string>{ "Korean" };
-            languages["Kyrgyz"] = new HashSet<string>{ "Kyrgyz" };
-            languages["Lao"] = new HashSet<string>{ "Lao" };
-            languages["Latvian"] = new HashSet<string>{ "Latvian" };
-            languages["Lithuanian"] = new HashSet<string>{ "Lithuanian" };
-            languages["Macedonian"] = new HashSet<string>{ "Macedonian" };
-            languages["Malay"] = new HashSet<string>{ "Malay" };
-            languages["Mandarin"] = new HashSet<string>{ "Mandarin" };
-            languages["Norwegian"] = new HashSet<string>{ "Norwegian" };
-            languages["Pashto"] = new HashSet<string>{ "Pashto" };
-            languages["Persian"] = new HashSet<string>{ "Persian" };
-            languages["Polish"] = new HashSet<string>{ "Polish" };
-            languages["Portuguese"] = new HashSet<string>{ "Portuguese" };
-            languages["Romanian"] = new HashSet<string>{ "Romanian" };
-            languages["Russian"] = new HashSet<string>{ "Russian", "RUssian" };
-            languages["Slovak"] = new HashSet<string>{ "Slovak" };
-            languages["Slovene"] = new HashSet<string>{ "Slovene", "Slovenian" };
-            languages["Somali"] = new HashSet<string>{ "Somali" };
-            languages["spanish"] = new HashSet<string>{ "Span", "Spanish" };
-            languages["Swahili"] = new HashSet<string>{ "Swahili" };
-            languages["Swedish"] = new HashSet<string>{ "Swedish" };
-            languages["Tagalog"] = new HashSet<string>{ "Tagalog" };
-            languages["Tajik"] = new HashSet<string>{ "Tajik" };
-            languages["Tamil"] = new HashSet<string>{ "Tamil" };
-            languages["Thai"] = new HashSet<string>{ "Thai" };
-            languages["Tigrignya"] = new HashSet<string>{ "Tigrigna", "Tigrignya", "Tigrinya" };
-            languages["Turkish"] = new HashSet<string>{ "Turkish" };
-            languages["Ukrainian"] = new HashSet<string>{ "Ukrainian" };
-            languages["Urdu"] = new HashSet<string>{ "Urdu" };
-            languages["Vietnamese"] = new HashSet<string>{ "Vietnamese" };
-            languages["Xhosa"] = new HashSet<string>{ "Xhosa" };
+            languages["Afrikaans"] = new HashSet<string> { "Afrikaans" };
+            languages["Albanian"] = new HashSet<string> { "Albanian" };
+            languages["Amharic"] = new HashSet<string> { "Amharic" };
+            languages["Arabic"] = new HashSet<string> { "Arabic", "Arabi" };
+            languages["Armenian"] = new HashSet<string> { "Armenian" };
+            languages["Azeri"] = new HashSet<string> { "Azeri" };
+            languages["Belarusian"] = new HashSet<string> { "Belarusian" };
+            languages["Bengali"] = new HashSet<string> { "Bengali" };
+            languages["Bulgarian"] = new HashSet<string> { "Bulgarian" };
+            languages["Burmese"] = new HashSet<string> { "Burmese" };
+            languages["Cambodian"] = new HashSet<string> { "Cambodian" };
+            languages["Cantonese"] = new HashSet<string> { "Cantonese" };
+            languages["Chinese"] = new HashSet<string> { "Chinese" };
+            languages["Creole"] = new HashSet<string> { "Creole" };
+            languages["Czech"] = new HashSet<string> { "Czech" };
+            languages["Danish"] = new HashSet<string> { "Danish" };
+            languages["Dari"] = new HashSet<string> { "Dari" };
+            languages["Dutch"] = new HashSet<string> { "Dutch" };
+            languages["English"] = new HashSet<string> { "Eng", "English", "Enlgish" };
+            languages["Estonian"] = new HashSet<string> { "Estonian" };
+            languages["Finnish"] = new HashSet<string> { "Finnish" };
+            languages["French"] = new HashSet<string> { "French" };
+            languages["Georgian"] = new HashSet<string> { "Georgian" };
+            languages["German"] = new HashSet<string> { "German" };
+            languages["Greek"] = new HashSet<string> { "Greek" };
+            languages["Hebrew"] = new HashSet<string> { "Hebrew" };
+            languages["Hindi"] = new HashSet<string> { "Hindi" };
+            languages["Hungarian"] = new HashSet<string> { "Hungarian" };
+            languages["Indonesian"] = new HashSet<string> { "Indonesian" };
+            languages["International"] = new HashSet<string> { "International" };
+            languages["Italian"] = new HashSet<string> { "Italian" };
+            languages["Japanese"] = new HashSet<string> { "Japanese" };
+            languages["Kazakh"] = new HashSet<string> { "Kazakh" };
+            languages["Kinyarwanda"] = new HashSet<string> { "Kinyarwanda" };
+            languages["Kirundi"] = new HashSet<string> { "Kirundi" };
+            languages["Korean"] = new HashSet<string> { "Korean" };
+            languages["Kyrgyz"] = new HashSet<string> { "Kyrgyz" };
+            languages["Lao"] = new HashSet<string> { "Lao" };
+            languages["Latvian"] = new HashSet<string> { "Latvian" };
+            languages["Lithuanian"] = new HashSet<string> { "Lithuanian" };
+            languages["Macedonian"] = new HashSet<string> { "Macedonian" };
+            languages["Malay"] = new HashSet<string> { "Malay" };
+            languages["Mandarin"] = new HashSet<string> { "Mandarin" };
+            languages["Norwegian"] = new HashSet<string> { "Norwegian" };
+            languages["Pashto"] = new HashSet<string> { "Pashto" };
+            languages["Persian"] = new HashSet<string> { "Persian" };
+            languages["Polish"] = new HashSet<string> { "Polish" };
+            languages["Portuguese"] = new HashSet<string> { "Portuguese" };
+            languages["Romanian"] = new HashSet<string> { "Romanian" };
+            languages["Russian"] = new HashSet<string> { "Russian", "RUssian" };
+            languages["Slovak"] = new HashSet<string> { "Slovak" };
+            languages["Slovene"] = new HashSet<string> { "Slovene", "Slovenian" };
+            languages["Somali"] = new HashSet<string> { "Somali" };
+            languages["spanish"] = new HashSet<string> { "Span", "Spanish" };
+            languages["Swahili"] = new HashSet<string> { "Swahili" };
+            languages["Swedish"] = new HashSet<string> { "Swedish" };
+            languages["Tagalog"] = new HashSet<string> { "Tagalog" };
+            languages["Tajik"] = new HashSet<string> { "Tajik" };
+            languages["Tamil"] = new HashSet<string> { "Tamil" };
+            languages["Thai"] = new HashSet<string> { "Thai" };
+            languages["Tigrignya"] = new HashSet<string> { "Tigrigna", "Tigrignya", "Tigrinya" };
+            languages["Turkish"] = new HashSet<string> { "Turkish" };
+            languages["Ukrainian"] = new HashSet<string> { "Ukrainian" };
+            languages["Urdu"] = new HashSet<string> { "Urdu" };
+            languages["Vietnamese"] = new HashSet<string> { "Vietnamese" };
+            languages["Xhosa"] = new HashSet<string> { "Xhosa" };
             languages["Zulu"] = new HashSet<string> { "Zulu" };
         }
         private void InitLanguageSelected()
@@ -327,21 +327,24 @@ namespace GUI
         /// </summary>
         private void Delete_Files()
         {
-            string target;
-            if (dest[dest.Length - 1] == '\\')
-                target = dest;
-            else target = dest + '\\';
-            DirectoryInfo d = new DirectoryInfo(target);
-            FileInfo[] Files = d.GetFiles();
-            foreach (FileInfo file in Files)
+            if(dest != "")
             {
-                try
+                string target;
+                if (dest[dest.Length - 1] == '\\')
+                    target = dest;
+                else target = dest + '\\';
+                DirectoryInfo d = new DirectoryInfo(target);
+                FileInfo[] Files = d.GetFiles();
+                foreach (FileInfo file in Files)
                 {
-                    File.Delete(target + file.Name);
-                }
-                catch
-                {
+                    try
+                    {
+                        File.Delete(target + file.Name);
+                    }
+                    catch
+                    {
 
+                    }
                 }
             }
 
@@ -385,7 +388,7 @@ namespace GUI
                 stemming = true;
             else
                 stemming = false;
-            _postingFilesAPI = new PostingFilesManager(10, dest,stemming);
+            _postingFilesAPI = new PostingFilesManager(10, dest, stemming);
             idx = new Indexer(dest, Mode.Load, _postingFilesAPI);
             hasIndex = true;
             //call backgroundworker for the actual loading process
@@ -403,7 +406,7 @@ namespace GUI
 
         private void CheckIfLoadingSucceeded(object sender, RunWorkerCompletedEventArgs e)
         {
-            
+
             if (!noStemLoadSuccess && !stemLoadSuccess)
             {
                 System.Windows.MessageBox.Show("Dictionary loading failed.", "Loading Error", MessageBoxButton.OK, MessageBoxImage.Error);
@@ -423,7 +426,7 @@ namespace GUI
         {
             if (stemming)
             {
-                stemLoadSuccess =idx.LoadMainDictionaryFromMemory(stemming);
+                stemLoadSuccess = idx.LoadMainDictionaryFromMemory(stemming);
                 noStemLoadSuccess = false;
             }
             else
@@ -510,7 +513,7 @@ namespace GUI
             else
                 stemming = false;
 
-            if (mainTabControl.SelectedIndex == 1 && ( (stemming&&!stemLoadSuccess) || (!stemming && !noStemLoadSuccess)) )
+            if (mainTabControl.SelectedIndex == 1 && ((stemming && !stemLoadSuccess) || (!stemming && !noStemLoadSuccess)))
             {
                 System.Windows.MessageBox.Show("Please start the indexing or load a dictionary first.", "Path Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 // mainTabControl.SelectedIndex = 0;
@@ -577,7 +580,7 @@ namespace GUI
             Parser.InitStopWords(stopwords);
             if (!String.IsNullOrWhiteSpace(txtbxUserQuery.Text))
             {
-                SearchQuery(txtbxUserQuery.Text.Split(QuerrySplitters),"295");
+                SearchQuery(txtbxUserQuery.Text.Split(QuerrySplitters), "295");
             }
             if (!String.IsNullOrWhiteSpace(txtbxFileQuery.Text))
             {
@@ -622,7 +625,7 @@ namespace GUI
             Dictionary<string, int> termsInQuery = idx.ParseQuery(query, stemming);
             Dictionary<string, PostingFileRecord> releventDocuments = searcher.FindReleventDocuments(termsInQuery);
             HashSet<string> chosenLanguages = ExtractChosenLanguages();
-            rankedDocument = ranker.RankDocuments(termsInQuery, queryID, releventDocuments, rankedDocument,idx.AvgDocumentLength, chosenLanguages);
+            rankedDocument = ranker.RankDocuments(termsInQuery, queryID, releventDocuments, rankedDocument, idx.AvgDocumentLength, chosenLanguages);
         }
 
         private HashSet<string> ExtractChosenLanguages()
@@ -657,17 +660,32 @@ namespace GUI
             {
                 int resultsCount = rankedDocument.Length;
                 string[] queriesRecords = new string[resultsCount];
-                for(int i = 0; i< resultsCount; i++)
+                for (int i = 0; i < resultsCount; i++)
                 {
                     queriesRecords[i] = rankedDocument[i].ToString();
                 }
                 File.WriteAllLines(win.FileName, queriesRecords);
             }
-            
+
         }
 
         private void Test_Click(object sender, RoutedEventArgs e)
         {
+            try
+            {
+                Test();
+            }
+            catch
+            {
+                // Test_Click(sender, e);
+                Test();
+
+            }
+        }
+
+        void Test()
+        {
+
             if (String.IsNullOrWhiteSpace(txtbxUserQuery.Text) && String.IsNullOrWhiteSpace(txtbxFileQuery.Text))
             {
                 System.Windows.MessageBox.Show("Please submit or upload a query");
@@ -694,44 +712,49 @@ namespace GUI
                 stopwords = src + "stop_words.txt";
             else stopwords = src + "\\stop_words.txt";
             Parser.InitStopWords(stopwords);
-            for (double w=1; w<=2; w += 0.2)
+            for (double w1 = 2; w1 <= 2; w1 += 0.2)
             {
-                ranker.w2 = w;
-                rankedDocument = new DocumentRank[0];
-                SearchFileQuery();
-                int resultsCount = rankedDocument.Length;
-                string[] queriesRecords = new string[resultsCount];
-                for (int i = 0; i < resultsCount; i++)
+                for (double w2 = 0; w2 <= 2; w2 += 0.2)
                 {
-                    queriesRecords[i] = rankedDocument[i].ToString();
-                }
-                string fileName = String.Format("results.txt");
-                string filePath = String.Format(@"C:\IR\{0}", fileName);
-                File.WriteAllLines(filePath, queriesRecords);
-
-
-                Process p = new Process();
-                ProcessStartInfo info = new ProcessStartInfo();
-                info.FileName = "cmd.exe";
-                info.RedirectStandardInput = true;
-                info.UseShellExecute = false;
-
-                p.StartInfo = info;
-                p.Start();
-
-                using (StreamWriter sw = p.StandardInput)
-                {
-                    if (sw.BaseStream.CanWrite)
+                    ranker.w1 = w1;
+                    ranker.w2 = w2;
+                    rankedDocument = new DocumentRank[0];
+                    SearchFileQuery();
+                    int resultsCount = rankedDocument.Length;
+                    string[] queriesRecords = new string[resultsCount];
+                    for (int i = 0; i < resultsCount; i++)
                     {
-                        sw.WriteLine(@"cd C:\IR");
-                        sw.WriteLine(String.Format("treceval qrels.txt {0} > output_{0}", fileName));
+                        queriesRecords[i] = rankedDocument[i].ToString();
                     }
-                }
+                    string fileName = String.Format("results_w1_{0}_w2_{1}.txt", w1,w2);
+                    string filePath = String.Format(@"C:\IR\{0}", fileName);
+                    File.WriteAllLines(filePath, queriesRecords);
 
+
+                    Process p = new Process();
+                    ProcessStartInfo info = new ProcessStartInfo();
+                    info.FileName = "cmd.exe";
+                    info.RedirectStandardInput = true;
+                    info.UseShellExecute = false;
+
+                    p.StartInfo = info;
+                    p.Start();
+
+                    using (StreamWriter sw = p.StandardInput)
+                    {
+                        if (sw.BaseStream.CanWrite)
+                        {
+                            sw.WriteLine(@"cd C:\IR");
+                            sw.WriteLine(String.Format("treceval qrels.txt {0} > output_{0}", fileName));
+                        }
+                    }
+
+                }
 
             }
             System.Windows.MessageBox.Show("Test is done");
-
+            return;
         }
     }
+
 }
